requirements.txt
python-telegram-bot==20.7
aiohttp==3.8.5
pandas==2.2.3
numpy==1.26.4
python-dotenv
python-telegram-bot==20.6
ccxt==4.1.89
pandas==2.2.1
matplotlib==3.8.2
ta==0.10.2
numpy==1.26.4
nest_asyncio==1.6.0
